+++ wxc_sdk/telephony/prem_pstn/trunk.py
@@ -5,13 +5,13 @@ from typing import List, Any, Optional
 from pydantic import Field, TypeAdapter
 
 from ...api_child import ApiChild
-from ...base import SafeEnum as Enum
+from ...base import SafeEnum as Enum, enum_str
 from ...base import to_camel, ApiModel
 from ...common import Customer, IdAndName
 
 __all__ = ['TrunkType', 'Trunk', 'TrunkDeviceType', 'TrunkTypeWithDeviceType', 'DeviceStatus',
            'ResponseStatusType', 'ResponseStatus', 'CnameRecord', 'OutboundProxy', 'TrunkDetail', 'TrunkUsage',
-           'TrunkApi']
+           'PChargeInfoSupportPolicy', 'TrunkApi']
 
 
 class TrunkType(str, Enum):
@@ -87,6 +87,18 @@ class OutboundProxy(ApiModel):
     cname_records: Optional[list[CnameRecord]] = None
     attachment_updated: bool
 
+class PChargeInfoSupportPolicy(str, Enum):
+    #: The P-Charge-Info header support policy is disabled.
+    disabled = 'DISABLED'
+    #: The P-Charge-Info header is always included in outbound PSTN calls using Webex Calling primary number or
+    #: location’s main number.
+    asserted_identity = 'ASSERTED_IDENTITY'
+    #: The P-Charge-Info header is included in outbound PSTN calls using the originating or redirecting Webex Calling
+    #: entity's location charge number if set, else the entity's primary number if set and not toll-free, else the
+    #: main number of the entity's location if set and not toll-free. If none of these are set or not toll-free, it
+    #: uses the same number as the ASSERTED_IDENTITY option.
+    configurable_charge_number = 'CONFIGURABLE_CHARGE_NUMBER'
+
 
 class TrunkDetail(ApiModel):
     trunk_id: str = Field(alias='id')
@@ -130,6 +142,7 @@ class TrunkDetail(ApiModel):
     max_concurrent_calls: int
     #: Flag to indicate if the trunk is restricted to a dedicated instance.
     is_restricted_to_dedicated_instance: Optional[bool] = False
+    p_charge_info_support_policy: Optional[PChargeInfoSupportPolicy] = None
 
 
 class TrunkUsage(ApiModel):
@@ -181,7 +194,8 @@ class TrunkApi(ApiChild, base='telephony/config/premisePstn/trunks'):
 
     def create(self, name: str, location_id: str, password: str, trunk_type: TrunkType = TrunkType.registering,
                dual_identity_support_enabled: bool = None, device_type: TrunkDeviceType = None, address: str = None,
-               domain: str = None, port: int = None, max_concurrent_calls: int = None, org_id: str = None) -> str:
+               domain: str = None, port: int = None, max_concurrent_calls: int = None,
+               p_charge_info_support_policy: PChargeInfoSupportPolicy = None, org_id: str = None) -> str:
         """
         Create a Trunk for the organization.
 
@@ -212,14 +226,33 @@ class TrunkApi(ApiChild, base='telephony/config/premisePstn/trunks'):
         :type port: int
         :param max_concurrent_calls: Max Concurrent call. Required to create a static certificate based trunk.
         :type max_concurrent_calls: int
+        :param p_charge_info_support_policy: P-Charge Info Support policy.
+        :type p_charge_info_support_policy: PChargeInfoSupportPolicy
         :param org_id: Organization to which trunk belongs.
         :type org_id: str
         :return: id of new trunk
         :rtype: str
         """
-        body = {to_camel(p): v for p, v in locals().items()
-                if p not in {'self', 'org_id'} and v is not None}
         params = org_id and {'orgId': org_id} or None
+        body = dict()
+        body['name'] = name
+        body['locationId'] = location_id
+        body['password'] = password
+        if dual_identity_support_enabled is not None:
+            body['dualIdentitySupportEnabled'] = dual_identity_support_enabled
+        body['trunkType'] = enum_str(trunk_type)
+        if device_type is not None:
+            body['deviceType'] = device_type
+        if address is not None:
+            body['address'] = address
+        if domain is not None:
+            body['domain'] = domain
+        if port is not None:
+            body['port'] = port
+        if max_concurrent_calls is not None:
+            body['maxConcurrentCalls'] = max_concurrent_calls
+        if p_charge_info_support_policy is not None:
+            body['pChargeInfoSupportPolicy'] = enum_str(p_charge_info_support_policy)
         url = self.ep()
         data = self.post(url=url, params=params, json=body)
         return data['id']
@@ -248,7 +281,8 @@ class TrunkApi(ApiChild, base='telephony/config/premisePstn/trunks'):
         return TrunkDetail.model_validate(data)
 
     def update(self, trunk_id: str, name: str, password: str, dual_identity_support_enabled: bool = None,
-               max_concurrent_calls: int = None, org_id: str = None):
+               max_concurrent_calls: int = None, p_charge_info_support_policy: PChargeInfoSupportPolicy = None,
+               org_id: str = None):
         """
         Modify a Trunk for the organization.
 
@@ -267,6 +301,8 @@ class TrunkApi(ApiChild, base='telephony/config/premisePstn/trunks'):
         :type dual_identity_support_enabled: bool
         :param max_concurrent_calls: Max Concurrent call. Required to create a static certificate based trunk.
         :type max_concurrent_calls: int
+        :param p_charge_info_support_policy: P-Charge Info Support policy.
+        :type p_charge_info_support_policy: PChargeInfoSupportPolicy
         :param org_id: Organization to which trunk belongs.
         :type org_id: str:return:
         """
@@ -280,6 +316,8 @@ class TrunkApi(ApiChild, base='telephony/config/premisePstn/trunks'):
             body['dualIdentitySupportEnabled'] = dual_identity_support_enabled
         if max_concurrent_calls is not None:
             body['maxConcurrentCalls'] = max_concurrent_calls
+        if p_charge_info_support_policy is not None:
+            body['pChargeInfoSupportPolicy'] = enum_str(p_charge_info_support_policy)
         url = self.ep(trunk_id)
         self.put(url=url, params=params, json=body)