+++ docs/user/changes.rst
@@ -1,6 +1,8 @@
 Release history
 ===============
 
+- feat: support for pChargeInfoSupportPolicy on :class:`trunk details <wxc_sdk.telephony.prem_pstn.trunk.TrunkDetail>` and in :meth:`api.telephony.prem_pstn.trunk.create <wxc_sdk.telephony.prem_pstn.trunk.TrunkApi.create>` and :meth:`api.telephony.prem_pstn.trunk.update <wxc_sdk.telephony.prem_pstn.trunk.TrunkApi.update>`
+
 1.26.0
 ------
 - feat: new API :attr:`api.telephony.hotdesking_voiceportal <wxc_sdk.telephony.TelephonyApi.hotdesking_voiceportal>`